refactor : 서비스 계층 분리 (CQRS 패턴)
