+++ src/main/java/site/codemonster/comon/domain/problem/controller/AdminProblemApiController.java
@@ -12,7 +12,8 @@ import site.codemonster.comon.domain.problem.dto.request.ProblemInfoRequest;
 import site.codemonster.comon.domain.problem.dto.response.ProblemInfoResponse;
 import site.codemonster.comon.domain.problem.entity.Problem;
 import site.codemonster.comon.domain.problem.enums.Platform;
-import site.codemonster.comon.domain.problem.service.ProblemService;
+import site.codemonster.comon.domain.problem.service.ProblemCommandService;
+import site.codemonster.comon.domain.problem.service.ProblemQueryService;
 import site.codemonster.comon.global.error.dto.response.ApiResponse;
 
 import java.util.List;
@@ -23,11 +24,12 @@ import java.util.Map;
 @RequiredArgsConstructor
 public class AdminProblemApiController {
 
-    private final ProblemService problemService;
+    private final ProblemCommandService problemCommandService;
+    private final ProblemQueryService problemQueryService;
 
     @PostMapping("/check/baekjoon")
     public ResponseEntity<?> checkBaekjoonProblem(@RequestParam String problemId) {
-        ProblemInfoResponse response = problemService.checkProblem(problemId, Platform.BAEKJOON);
+        ProblemInfoResponse response = problemCommandService.checkProblem(problemId, Platform.BAEKJOON);
 
         return ResponseEntity.status(PROBLEM_CHECK_SUCCESS.getStatusCode())
                 .contentType(MediaType.APPLICATION_JSON)
@@ -38,7 +40,7 @@ public class AdminProblemApiController {
     public ResponseEntity<?> checkProgrammersProblem(@RequestBody @Valid ProblemInfoRequest request) {
         if (request.getPlatform() == null) request.setPlatform(Platform.PROGRAMMERS);
 
-        ProblemInfoResponse response = problemService.checkProblem(request);
+        ProblemInfoResponse response = problemCommandService.checkProblem(request);
 
         return ResponseEntity.status(PROBLEM_CHECK_SUCCESS.getStatusCode())
                 .contentType(MediaType.APPLICATION_JSON)
@@ -47,7 +49,7 @@ public class AdminProblemApiController {
 
     @PostMapping("/check/leetcode")
     public ResponseEntity<?> checkLeetcodeProblem(@RequestParam String url) {
-        ProblemInfoResponse response = problemService.checkProblem(url, Platform.LEETCODE);
+        ProblemInfoResponse response = problemCommandService.checkProblem(url, Platform.LEETCODE);
 
         return ResponseEntity.status(PROBLEM_CHECK_SUCCESS.getStatusCode())
                 .contentType(MediaType.APPLICATION_JSON)
@@ -56,7 +58,7 @@ public class AdminProblemApiController {
 
     @PostMapping("/register/batch")
     public ResponseEntity<?> registerProblems(@RequestBody @Valid ProblemBatchRequest batchRequest) {
-        List<Problem> savedProblems = problemService.registerProblems(batchRequest.getProblems());
+        List<Problem> savedProblems = problemCommandService.registerProblems(batchRequest.getProblems());
 
         Map<String, Object> responseData = Map.of(
                 "count", savedProblems.size(),
@@ -70,7 +72,7 @@ public class AdminProblemApiController {
 
     @GetMapping("/statistics")
     public ResponseEntity<?> getProblemStatistics() {
-        Map<Platform, Long> statistics = problemService.getProblemStatistics();
+        Map<Platform, Long> statistics = problemQueryService.getProblemStatistics();
 
         return ResponseEntity.status(PROBLEM_STATISTICS_GET_SUCCESS.getStatusCode())
                 .contentType(MediaType.APPLICATION_JSON)
@@ -79,7 +81,7 @@ public class AdminProblemApiController {
 
     @GetMapping("/api/list")
     public ResponseEntity<?> getProblemList() {
-        List<Problem> problems = problemService.getAllProblems();
+        List<Problem> problems = problemQueryService.getAllProblems();
 
         return ResponseEntity.status(PROBLEM_LIST_GET_SUCCESS.getStatusCode())
                 .contentType(MediaType.APPLICATION_JSON)
@@ -88,7 +90,7 @@ public class AdminProblemApiController {
 
     @GetMapping("/api/list-by-platform")
     public ResponseEntity<?> getProblemListByPlatform(@RequestParam(required = false) String platform) {
-        List<Problem> problems = problemService.getProblemsByPlatform(Platform.valueOf(platform.toUpperCase()));
+        List<Problem> problems = problemQueryService.getProblemsByPlatform(Platform.valueOf(platform.toUpperCase()));
 
         return ResponseEntity.status(PROBLEM_LIST_GET_SUCCESS.getStatusCode())
                 .contentType(MediaType.APPLICATION_JSON)
@@ -97,7 +99,7 @@ public class AdminProblemApiController {
 
     @PutMapping("/api/{problemId}")
     public ResponseEntity<?> updateProblem(@PathVariable Long problemId, @RequestBody Map<String, String> updateData) {
-        Problem updatedProblem = problemService.updateProblem(problemId, updateData);
+        Problem updatedProblem = problemCommandService.updateProblem(problemId, updateData);
 
         return ResponseEntity.status(PROBLEM_UPDATE_SUCCESS.getStatusCode())
                 .contentType(MediaType.APPLICATION_JSON)
@@ -106,7 +108,7 @@ public class AdminProblemApiController {
 
     @DeleteMapping("/api/{problemId}")
     public ResponseEntity<?> deleteProblem(@PathVariable Long problemId) {
-        problemService.deleteProblem(problemId);
+        problemCommandService.deleteProblem(problemId);
 
         return ResponseEntity.status(PROBLEM_DELETE_SUCCESS.getStatusCode())
                 .contentType(MediaType.APPLICATION_JSON)
